Collected small changes and fixes for the next feature release

+++ src/INTEL/npair_full_bin_ghost_intel.cpp
@@ -559,7 +559,7 @@ void NPairFullBinGhostIntel::fbi(const int offload, NeighList * list,
         if (ct + obound > list_size) {
           if (i < ito - 1) {
             *overflow = 1;
-            ct = (ifrom + tid * 2) * maxnbors;
+            ct = (bigint)(ifrom + tid * 2) * (bigint)maxnbors;
           }
         }
       }
